add styled button component/rename folder forms to form
