+++ components/Button/Button.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export default function AddButton() {
+  return <StyledButton>hello</StyledButton>;
+}
+
+const StyledButton = styled.button`
+  padding: 0.2em 0.5em;
+`;
